# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ERP System API — PR/PO/Store Management with full Approval workflow",
        "title": "ERP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ERP Support",
            "email": "support@erp.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/matcode/brands": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "List brands",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Create or update brand (UPSERT)",
                "parameters": [
                    {
                        "description": "Brand data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBrandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Brand"
                        }
                    }
                }
            }
        },
        "/matcode/brands/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Get brand by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Brand"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Delete brand",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/matcode/codes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "List all material codes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Create material code (auto-generates mat_code)",
                "parameters": [
                    {
                        "description": "Material code data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMaterialCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.MaterialCode"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/matcode/codes/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Get material code by mat_code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mat Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaterialCode"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Update material code (is_active only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mat Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMaterialCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Delete material code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mat Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/matcode/matnames": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "List material names (filter by subgroup_code)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subgroup_code filter",
                        "name": "subgroup",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Create or update material name (UPSERT)",
                "parameters": [
                    {
                        "description": "MatName data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMatNameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MatName"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/matcode/matnames/{subgroup}/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Get material name by subgroup_code and mat_name_code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subgroup Code",
                        "name": "subgroup",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MatName Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MatName"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaterialCode"
                ],
                "summary": "Delete material name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subgroup Code",
                        "name": "subgroup",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MatName Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
...
        "models.Brand": {
            "type": "object",
            "properties": {
                "brand_code": {
                    "type": "string"
                },
                "brand_name": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateBrandRequest": {
            "type": "object",
            "required": [
                "brand_code",
                "brand_name"
            ],
            "properties": {
                "brand_code": {
                    "type": "string",
                    "maxLength": 20
                },
                "brand_name": {
                    "type": "string",
                    "maxLength": 200
                }
            }
        },
...
        "models.MatName": {
            "type": "object",
            "properties": {
                "mat_name_code": {
                    "type": "string"
                },
                "mat_name_en": {
                    "type": "string"
                },
                "mat_name_th": {
                    "type": "string"
                },
                "subgroup_code": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateMaterialCodeRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}